style(import): [table] display pagination at the bottom always visible
